Método para actualizar en el service

+++ src/main/java/co/edu/usbcali/aerolinea/dtos/AsientoDTO.java
@@ -16,7 +16,7 @@ public class AsientoDTO {
     private String ModeloAvion;
     private Integer avion_avioId;
     private String ubicacion;
-    private float precio;
+    private Integer precio;
     private  String estado;
 }
